Extract discover items for cleanliness
