+++ site/gatsby-site/src/components/discover/parseURL.js
@@ -70,7 +70,5 @@ export default function ({ location, indexName, queryConfig }) {
 
   const searchState = generateSearchState({ query });
 
-  console.log('holis parseurl', searchState);
-
   return { [indexName]: searchState };
 }
